requirements.txt
pyperclip==1.8.2
python-slugify==8.0.4
ttkbootstrap==1.10.1
